actualizo salir grupo
El creador no puedo salir del grupo

+++ includes/DAO/DAOGrupos.php
@@ -175,13 +175,17 @@ class DAOGrupos extends DAO {
     }
 
     public function salirGrupo($idGrupo,$idUsuario){
+        $sql = 'SELECT * FROM grupos WHERE idUsuario = "' . $idUsuario . '" AND idGrupo = "' . $idGrupo. '"';
+        $consulta = $this->ejecutarConsulta($sql);
 
-
-        $query = 'DELETE FROM `grupos-usuarios` WHERE idUsuario = "' . $idUsuario . '" AND idGrupo = "' . $idGrupo. '"';
-        $resultado = $this->ejecutarModificacion($query);
-
-        if($resultado == 0){
+        if(count($consulta) == 0)
+            $query = 'DELETE FROM `grupos-usuarios` WHERE idUsuario = "' . $idUsuario . '" AND idGrupo = "' . $idGrupo. '"';
+            $resultado = $this->ejecutarModificacion($query);
+            if($resultado == 0){
             return "Error al salir del grupo";
+        }
+        else{
+            return "Eres el creador del grupo";
         }
 
     }
